+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 - [New] QuoteSource supports getting data in parallel when `num_workers` > 1.
+- [New] Add performance tearsheet for IB account.
 
 ## 0.1.33
 - [Fix] Fix process status display on `Home` page.